Split the sources of orbital.rs
